Conflicts:
	README.md

+++ encFile.py
@@ -1,7 +1,9 @@
 import os
+
 try:
     import pwd
-except: ImportError
+except ImportError:
+    pass
 
 from Crypto.Hash import SHA256
 from Crypto.Cipher import AES
@@ -26,20 +28,46 @@ encrypted_key = ""
 def check_file_creation():
     global os_version
     path = os.getcwd()
+    os_version = get_os()
     # Check for windows
     if path.find(":\\") != -1:
-        openfile = open(path + "\\" + storage_file, "wb")
-        openfile.close()
+        if not os.path.isfile(storage_file):
+            openfile = open(path + "\\" + storage_file, "wb")
+            id = id_create()
+            openfile.write(id)
+            openfile.write("\n")
+            key = create_key()
+            openfile.write(key)
+            openfile.write("\n")
+            openfile.close()
         os_version = 1
     # Linux
     else:
-        openfile = open("./" + storage_file, "wb")
-        openfile.close()
+        if not os.path.isfile(storage_file):
+            openfile = open("./" + storage_file, "wb")
+            id = id_create()
+            openfile.write(id)
+            openfile.write("\n")
+            key = create_key()
+            openfile.write(key)
+            openfile.write("\n")
+            openfile.close()
+
+
+# Get OS Type
+def get_os():
+    global os_version
+    path = os.getcwd()
+    # Check for windows
+    if path.find(":\\") != -1:
+        os_version = 1
+    return os_version
 
 
 # Get username
 def get_username():
     global username
+    get_os()
     # Windows
     if os_version == 1:
         path = os.path.join(os.path.expandvars("%userprofile%"), "Documents and Settings")
@@ -48,6 +76,7 @@ def get_username():
     # Linux
     else:
         username = pwd.getpwuid(os.getuid()).pw_name
+    return username
 
 
 # Verify that the file is correct by submitting the verify id from the file
@@ -56,7 +85,6 @@ def verify(verify_id):
     if not result:
         print "ERROR: INVALID FILE"
         exit()
-
 
 
 # Verify unique ID
@@ -69,8 +97,13 @@ def id_create():
     get_username()
     newhash = SHA256.new()
     newhash.update(username + os.getcwd())
-    newid = newhash.digest
+    newid = newhash.digest()
     return newid
+
+
+# Debugger print function
+def debugger():
+    print username
 
 
 # Store encrypted key as global var
@@ -79,10 +112,12 @@ def current_key(currentkey):
     encrypted_key = currentkey
 
 
-
 # Get key from provided line - recommended to store key in encrypted form until it is needed
 def obtain_key(line):
-    return decrypt(os.path, line)
+    key = os.getcwd() + get_username()
+    while len(key) < 32:
+        key += os.getcwd() + get_username()
+    return decrypt(key[:32], line)
 
 
 # Create key for file - temporary key generation method
@@ -90,8 +125,11 @@ def create_key():
     # Python time code from http://stackoverflow.com/questions/35318841/python-how-to-get-location-time-in-windows
     utc_offset = time.strftime('%z')
     tz_name = time.tzname[0]
-    current_key(encrypt(os.getcwd(), get_username() + os.getcwd() + tz_name + utc_offset))
-    return encrypt(os.getcwd(), get_username() + os.getcwd() + tz_name + utc_offset) # May not be needed
+    key = os.getcwd() + get_username()
+    while len(key) < 32:
+        key += os.getcwd() + get_username()
+    current_key(encrypt(key[:32], username + os.getcwd()))
+    return encrypt(key[:32], username + os.getcwd())  # May not be needed
 
 
 # Encryption from http://stackoverflow.com/questions/12524994/encrypt-decrypt-using-pycrypto-aes-256
@@ -105,14 +143,17 @@ def pad(s):
 
 # Unpadding
 def unpad(s):
-    return s[:-ord(s[len(s)-1:])]
+    return s[:-ord(s[len(s) - 1:])]
 
 
 # Encrypt information
 def encrypt(key, raw):
     raw = pad(raw)
+    key1 = key[:32]
+    key2 = key1.encode("ASCII", 'ignore')
+    key3 = key2[:32]
     iv = Random.new().read(AES.block_size)
-    cipher = AES.new(key, AES.MODE_CBC, iv)
+    cipher = AES.new(key3, AES.MODE_CBC, iv)
     return base64.b64encode(iv + cipher.encrypt(raw))
 
 
@@ -121,6 +162,9 @@ def decrypt(key, enc):
     if enc is None:
         return ""
     enc = base64.b64decode(enc)
+    key1 = key[:32]
+    key2 = key1.encode("ASCII", 'ignore')
+    key3 = key2[:32]
     iv = enc[:16]
-    cipher = AES.new(key, AES.MODE_CBC, iv)
+    cipher = AES.new(key3, AES.MODE_CBC, iv)
     return unpad(cipher.decrypt(enc[16:]))
